feat(models): Add SimpleProject model for project details
