+++ analyze_git_projects/models/__init__.py
@@ -1,0 +1,2 @@
+from .simple_project_schema import SimpleProject
+from .other_models import RepositoryInfo  # Only keep RepositoryInfo for URL parsing
